:recycle: Simplify tab route code.
Move management mode detection to the parent component.

I'm starting to wonder if I should be using a flag instead?
e.g.
/datasets/:slug/visualize?mode=management

with a v-switch/toggle to change the mode?
