+++ web/src/components/dataset-layout/AccessTab.vue
@@ -1,20 +1,17 @@
 <template>
-  <LockedTab
-    v-if="locked || !isManaging"
-    :tooltip-text="tooltipText"
-    >Access</LockedTab
-  >
   <v-tab
-    v-else
+    v-if="isManaging"
     :to="{ name: 'DatasetAccessManagePage', params: { slug } }"
     ><h3>Access</h3></v-tab
+  >
+  <LockedTab
+    v-else
+    tooltip-text="Cannot access this tab unless in edit mode."
+    >Access</LockedTab
   >
 </template>
 
 <script lang="ts" setup>
-import { computed } from "vue"
-import { useRoute } from "vue-router"
-
 import LockedTab from "@/components/LockedTab.vue"
 
 defineProps({
@@ -22,29 +19,9 @@ defineProps({
     type: String,
     required: true,
   },
-  locked: {
+  isManaging: {
     type: Boolean,
     default: false,
   },
-})
-
-const route = useRoute()
-
-const isManaging = computed(() => {
-  switch (route.name) {
-    case "DatasetDescriptionManagePage":
-    case "DatasetFieldsManagePage":
-    case "DatasetAccessManagePage":
-    case "DatasetVisualizeManagePage":
-      return true
-    default:
-      return false
-  }
-})
-
-const tooltipText = computed<string | undefined>(() => {
-  if (isManaging.value) return
-
-  return "Cannot access this tab unless in edit mode."
 })
 </script>
