+++ web/src/components/dataset-layout/FieldsTab.vue
@@ -1,41 +1,24 @@
 <template>
-  <LockedTab v-if="locked">Fields</LockedTab>
-  <v-tab
-    v-else
-    :to="{ name: routeName, params: { slug } }"
-    ><h3>Fields</h3></v-tab
-  >
+  <v-tab :to="{ name: routeName, params: { slug } }"><h3>Fields</h3></v-tab>
 </template>
 
 <script lang="ts" setup>
 import { computed } from "vue"
-import { useRoute } from "vue-router"
 
-import LockedTab from "@/components/LockedTab.vue"
-
-defineProps({
+const props = defineProps({
   slug: {
     type: String,
     required: true,
   },
-  locked: {
+  isManaging: {
     type: Boolean,
     default: false,
   },
 })
 
-const route = useRoute()
-
-// TODO: maybe vary route by user role?
 const routeName = computed(() => {
-  switch (route.name) {
-    case "DatasetDescriptionManagePage":
-    case "DatasetFieldsManagePage":
-    case "DatasetAccessManagePage":
-    case "DatasetVisualizeManagePage":
-      return "DatasetFieldsManagePage"
-    default:
-      return "DatasetFieldsReadPage"
-  }
+  if (props.isManaging) return "DatasetFieldsManagePage"
+
+  return "DatasetFieldsReadPage"
 })
 </script>
